Reorganisation, cleanup.
