+++ xde/src/lib.rs
@@ -43,7 +43,6 @@ use illumos_sys_hdrs::KM_SLEEP;
 pub mod dls;
 pub mod ip;
 pub mod mac;
-mod mac_sys;
 pub mod route;
 pub mod secpolicy;
 pub mod sys;
